feat: 프로필정보 연결, addLink 추가
