+++ src/components/SharedComponent/Footer.jsx
@@ -1,7 +1,7 @@
-import facebook from "../assets/facebook.svg";
-import twitter from "../assets/twitter.svg";
-import youtube from "../assets/youtube.svg";
-import instagram from "../assets/instagram.svg";
+import facebook from "../../assets/facebook.svg";
+import twitter from "../../assets/twitter.svg";
+import youtube from "../../assets/youtube.svg";
+import instagram from "../../assets/instagram.svg";
 
 function Footer() {
   return (
