+++ src/pages/Folder/App.css
@@ -1,5 +1,5 @@
-@import url(styles/Globals.css);
-@import url(styles/HomeStyles.css);
+@import url(../../styles/Globals.css);
+@import url(../../styles/HomeStyles.css);
 .gnb {
   display: flex;
   gap: 0.6rem;
@@ -35,6 +35,43 @@
   font-size: 16px;
 }
 
+.add-link-container {
+  position: relative;
+  width: 80rem;
+  margin: 6rem 32rem 9rem;
+}
+
+.addLink {
+  width: 100%;
+  padding-left: 5rem;
+  height: 6.9rem;
+  border-radius: 1.5rem;
+  border: 1px solid var(--primary-color);
+  color: var(--gray-60-color);
+  font-size: 1.6rem;
+}
+
+.add {
+  width: 8.1rem;
+  font-size: 1.4rem;
+  font-weight: 600;
+  padding: 1rem 1.6rem;
+}
+
+.add-link-container > button {
+  position: absolute;
+  top: 50%;
+  right: 2rem;
+  transform: translateY(-50%);
+}
+
+.add-link-container > img {
+  position: absolute;
+  top: 50%;
+  left: 2rem;
+  transform: translateY(-50%);
+}
+
 folder {
   background-color: var(--white-color);
   padding-top: 4rem;
@@ -68,6 +105,7 @@ folder {
   border: none;
   border-radius: 1rem;
   padding-left: 4rem;
+  margin: 4rem auto;
 }
 
 .img-container {
